compile expressions to LLVM IR
This commit introduces expression compilation for expressions.
Expressions are compiled right after initialization (when ExprState is
constructed) in ExecInitExpr. The expression tree is compiled into
a single function in LLVM IR, which is then optimized using LLVM
optimization passes. Compiled code substitutes `evalfunc` pointer
in ExprState so that the fact that the expression is compiled as opposed to
being interpreted becomes an implementation detail.