llvmQual: use runtime context for unsupported expression types
The expression context used at compile time is the context of the
PlanState node associated with the expression. However, it is not enough
for some other expression types (notably Vars) to work correctly. This
commit fixes the context of unsupported types to the runtime context
passed as an argument to EvalExpr.