llvmQual: add support for Materialize set function return mode
This commit adds support for SFRM_Materialize return mode, also, it
changes how "hasSetArg" is evaluated. Previously hasSetArg was evaluated
based on whether the arg function can return a set, not on whether it
actually returned a set. Now hasSetArg is an LLVM variable and is only
true if the arg function returned a set.

Also, when a set functions returns the "end" result it is possible that it's
args are set args and have not yet returned "end" result. In this case we
should loop back until both the function and its args return "end" results.